requirements.txt
alembic==1.7.7
cffi==1.15.0
click==8.1.2
cryptography==37.0.2
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
SQLAlchemy==1.4.35
uWSGI==2.0.20
Werkzeug==2.1.1
zipp==3.8.0
